requirements.txt
asgiref==3.2.7
certifi==2020.4.5.1
chardet==3.0.4
defusedxml==0.6.0
Django==3.0.5
django-allauth==0.41.0
django-appconf==1.0.4
django-countries==6.1.2
django-crispy-forms==1.9.0
django-imagekit==4.0.2
django-paypal==1.0.0
djangorestframework==3.11.0
idna==2.9
oauthlib==3.1.0
pilkit==2.0
Pillow==7.1.2
psycopg2-binary==2.8.5
python3-openid==3.1.0
pytz==2019.3
requests==2.23.0
requests-oauthlib==1.3.0
six==1.14.0
sqlparse==0.3.1
stripe==2.46.0
urllib3==1.25.9
